+++ container-engine-lib/lib/backend_impls/docker/object_attributes_provider/enclave_object_attributes_provider.go
@@ -41,7 +41,7 @@ type DockerEnclaveObjectAttributesProvider interface {
 	// ForNetworkingSidecarContainer(serviceGUIDSidecarAttachedTo string) (DockerObjectAttributes, error)
 	// ForFilesArtifactExpanderContainer(serviceGUID string, artifactId string) (DockerObjectAttributes, error)
 	// ForFilesArtifactExpansionVolume(serviceGUID string, artifactId string) (DockerObjectAttributes, error)
-	// ForModuleContainer(moduleID string, moduleGUID string, privatePortNum uint16) (DockerObjectAttributes, error)
+	ForModuleContainer(moduleID string, moduleGUID string, privateIpAddr net.IP, privatePortNum uint16) (DockerObjectAttributes, error)
 }
 
 // Private so it can't be instantiated
@@ -139,6 +139,42 @@ func (provider *dockerEnclaveObjectAttributesProviderImpl) ForApiContainer(
 	return objectAttributes, nil
 }
 
+func (provider *dockerEnclaveObjectAttributesProviderImpl) ForModuleContainer(
+	moduleID string,
+	moduleGUID string,
+	privatePortId string,
+	privatePortSpec *port_spec.PortSpec,
+) (DockerObjectAttributes, error) {
+	name, err := provider.getNameForEnclaveObject([]string{
+		moduleContainerNameFragment,
+		moduleGUID,
+	})
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred creating the module container name object")
+	}
+
+	usedPorts := map[string]*port_spec.PortSpec{
+		privatePortId: privatePortSpec,
+	}
+	serializedPortsSpec, err := port_spec_serializer.SerializePortSpecs(usedPorts)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred serializing the following module container ports object to a string for storing in the ports label: %+v", usedPorts)
+	}
+
+	labels, err := provider.getLabelsForEnclaveObjectWithIDAndGUID(moduleID, moduleGUID)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting the module labels using ID '%v' and GUID '%v'", moduleID, moduleGUID)
+	}
+	labels[label_key_consts.ContainerTypeLabelKey] = label_value_consts.ModuleContainerTypeLabelValue
+	labels[label_key_consts.PortSpecsLabelKey] = serializedPortsSpec
+
+	objectAttributes, err := newDockerObjectAttributesImpl(name, labels)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred while creating the ObjectAttributesImpl with the name '%s' and labels '%+v'", name, labels)
+	}
+
+	return objectAttributes, nil
+}
 
 // ====================================================================================================
 //                                      Private Helper Functions
@@ -167,19 +203,28 @@ func (provider *dockerEnclaveObjectAttributesProviderImpl) getLabelsForEnclaveOb
 	}
 }
 
-/*
-func (provider *dockerEnclaveObjectAttributesProviderImpl) getLabelsForEnclaveObjectWithGUID(guid string) map[*docker_label_key.DockerLabelKey]*docker_label_value.DockerLabelValue {
+func (provider *dockerEnclaveObjectAttributesProviderImpl) getLabelsForEnclaveObjectWithGUID(guid string) (map[*docker_label_key.DockerLabelKey]*docker_label_value.DockerLabelValue, error) {
 	labels := provider.getLabelsForEnclaveObject()
-	labels[label_key_consts.GUIDLabelKey] = guid
-	return labels
+	guidLabelValue, err := docker_label_value.CreateNewDockerLabelValue(guid)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred creating a Docker label value from GUID string '%v'", guid)
+	}
+	labels[label_key_consts.GUIDLabelKey] = guidLabelValue
+	return labels, nil
 }
 
-func (provider *dockerEnclaveObjectAttributesProviderImpl) getLabelsForEnclaveObjectWithIDAndGUID(id, guid string) map[*docker_label_key.DockerLabelKey]*docker_label_value.DockerLabelValue {
-	labels := provider.getLabelsForEnclaveObjectWithGUID(guid)
-	labels[label_key_consts.IDLabelKey] = id
-	return labels
+func (provider *dockerEnclaveObjectAttributesProviderImpl) getLabelsForEnclaveObjectWithIDAndGUID(id, guid string) (map[*docker_label_key.DockerLabelKey]*docker_label_value.DockerLabelValue, error) {
+	labels, err := provider.getLabelsForEnclaveObjectWithGUID(guid)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting the enclave object labels with GUID '%v'", guid)
+	}
+	idLabelValue, err := docker_label_value.CreateNewDockerLabelValue(id)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred creating a Docker label value from ID string '%v'", id)
+	}
+	labels[label_key_consts.IDLabelKey] = idLabelValue
+	return labels, nil
 }
-*/
 
 func getLabelKeyValuesAsStrings(labels map[*docker_label_key.DockerLabelKey]*docker_label_value.DockerLabelValue) map[string]string {
 	result := map[string]string{}